% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_river_centerline.R
\name{get_river_centerline}
\alias{get_river_centerline}
\title{Get the "centerline" of the longest river along the DEM}
\usage{
get_river_centerline(dem, quiet = TRUE)
}
\arguments{
\item{dem}{Either a SpatRaster (created via \code{\link[terra:rast]{terra::rast()}}) or an object
that \code{\link[terra:rast]{terra::rast()}} can read to create a SpatRaster.}

\item{quiet}{Boolean: should execution proceed "quietly", without messages
(\code{TRUE}) or should progress updates be posted during centerline download and
interpolation (\code{FALSE})?}
}
\value{
An sfc object (as created by \code{\link[sf:sfc]{sf::st_sfc()}}), with LINESTRING
geometry, representing the center of the river within the DEM area.
}
\description{
Get the "centerline" of the longest river along the DEM
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dem <- system.file("elevation.tiff", package = "rrrem")
centerline <- get_river_centerline(dem)
\dontshow{\}) # examplesIf}
}
