% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rem.R
\name{make_rem}
\alias{make_rem}
\title{Make a Relative Elevation Model (REM)}
\usage{
make_rem(dem, centerline = NULL, n_points = 1000, quiet = TRUE)
}
\arguments{
\item{dem}{Either a SpatRaster (created via \code{\link[terra:rast]{terra::rast()}}) or an object
that \code{\link[terra:rast]{terra::rast()}} can read to create a SpatRaster.}

\item{centerline}{Optionally, an sfc object (as created via \code{\link[sf:sfc]{sf::sfc()}})
representing the center of the target river to base the relative elevation
model on. If \code{NULL}, the default, the longest river in the bounding box with
a name in OpenStreetMap will be used instead.}

\item{n_points}{The number of sample points along the river to base
interpolation on. More points will make interpolation take longer.}

\item{quiet}{Boolean: should execution proceed "quietly", without messages
(\code{TRUE}) or should progress updates be posted during centerline download and
interpolation (\code{FALSE})?}
}
\value{
A SpatRaster object (as created by \code{\link[terra:rast]{terra::rast()}}), representing the
difference between elevations in \code{dem} and the interpolated elevation
}
\description{
This function creates a relative elevation model, or REM, based off an
input digital elevation model (DEM) and the trajectory of a river. Simple
inverse distance weighting is used to calculate "base" elevations at all
pixels within the DEM, which are then subtracted to produce the output REM.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
make_rem(system.file("elevation.tiff", package = "rrrem"))
\dontshow{\}) # examplesIf}
}
