% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_along_line.R
\name{points_along_line}
\alias{points_along_line}
\title{Create a set of number of evenly spaced points along a line geometry}
\usage{
points_along_line(line, n_points = 1000)
}
\arguments{
\item{line}{The line to create points along. Must be an sfc object (as
created by \code{\link[sf:sfc]{sf::st_sfc()}}) with LINESTRING geometry.}

\item{n_points}{The number of points along the line to create.}
}
\value{
An sfc object with POINT geometry of length \code{n_points}.
}
\description{
Create a set of number of evenly spaced points along a line geometry
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dem <- system.file("elevation.tiff", package = "rrrem")
centerline <- get_river_centerline(dem)
center_points <- points_along_line(centerline, n_points = n_points)
\dontshow{\}) # examplesIf}
}
