% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrrem-package.R
\docType{package}
\name{rrrem-package}
\alias{rrrem}
\alias{rrrem-package}
\title{rrrem: An R Port of the 'RiverREM' Python Package}
\description{
Build and visualize relative elevation models ('REM') representing elevation relative to a river centerline. This package is an R port of the existing Python 'RiverREM' package \url{https://github.com/klarrieu/RiverREM}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikemahoney218.github.io/rrrem/}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

}
\keyword{internal}
